New class for adapter and remove old convert tools
