+++ app/src/main/java/com/kylecorry/trail_sense/tools/convert/ui/FragmentToolConvert.kt
@@ -1,13 +1,10 @@
 package com.kylecorry.trail_sense.tools.convert.ui
 
-import android.content.Context
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentManager
-import androidx.fragment.app.FragmentStatePagerAdapter
 import androidx.viewpager.widget.ViewPager
 import com.kylecorry.trail_sense.R
 import com.kylecorry.trail_sense.databinding.FragmentToolConvertBinding
@@ -39,18 +36,5 @@ class FragmentToolConvert : BoundFragment<FragmentToolConvertBinding>() {
         container: ViewGroup?
     ): FragmentToolConvertBinding {
         return FragmentToolConvertBinding.inflate(layoutInflater, container, false)
-    }
-}
-
-class ConvertViewPagerAdapter(fm: FragmentManager, val convertTools: ArrayList<Fragment>, val convertToolsNames: ArrayList<Int>, val context: Context) : FragmentStatePagerAdapter(fm, BEHAVIOR_RESUME_ONLY_CURRENT_FRAGMENT) {
-
-    override fun getCount(): Int  = convertTools.size
-
-    override fun getItem(i: Int): Fragment {
-        return convertTools[i]
-    }
-
-    override fun getPageTitle(position: Int): CharSequence {
-        return context.getText(convertToolsNames[position])
     }
 }
